requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
stripe==2.55.2
urllib3==1.26.3
Python==3.8.5
django-filter==2.4.0
celery~=5.0.5
djangorestframework~=3.12.2